flask_app/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.28
pymysql==1.0.2
python-dotenv==0.19.0
plotly==5.3.1
pandas==1.3.4
python-dateutil==2.8.2
